Comments

+++ Gemm_CUDA.py
@@ -10,9 +10,10 @@ def cuda_gemm_kernel(A, B, C):
     i, j = cuda.grid(2)
     #Ensure that they stay within valid bounds
     if i < C.shape[0] and j < C.shape[1]:
+        #Initialize the resultant for each thread as 0
         C_value = 0.0
         for k in range(A.shape[1]):
-            #Operation
+            #Inner loop iteration
             C_value += A[i, k] * B[k, j]
         #Store in memory
         C[i, j] = C_value
@@ -81,8 +82,8 @@ def main():
     threadsperblock = (16, 16)
     num_runs = 10
     #Randomized initial matrices, numba CUDA works with numpy arrays
-    A = np.random.rand(m, n)
-    B = np.random.rand(n, k)
+    A = np.random.rand(m, n).astype(np.float32)
+    B = np.random.rand(n, k).astype(np.float32)
 
     #Run against the GPU
     start = time.time()
